+++ src/gr/aueb/cf/cf9/cha18/BankApp/DTO/AccountWithdrawDTO.java
@@ -1,4 +1,6 @@
 package gr.aueb.cf.cf9.cha18.BankApp.DTO;
 
-public record AccountWithdrawDTO(String iban, double balance) {
+import java.math.BigDecimal;
+
+public record AccountWithdrawDTO(String iban, BigDecimal balance) {
 }
